pyproject.toml
[project]
name = "history-rca"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.1",
    "drain3>=0.9.11",
    "google-adk>=1.23.0",
    "litellm>=1.81.4",
    "matplotlib>=3.10.8",
    "networkx>=3.6.1",
    "numpy>=2.4.1",
    "openai>=2.16.0",
    "pandas>=3.0.0",
    "pyarrow>=23.0.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
]
